# src/main/resources/swagger_ccp.yaml
swagger: '2.0'
info:
  description: Credit Card service
  version: 1.0.0
  title: Credit Card Provider App
  contact:
    name:
    email:
host: 'localhost:8080'
basePath: /ccp/v1
tags:
  - name: credit-card-data-controller
    description: Credit Card Data Controller
paths:
  /add-card:
    post:
      tags:
        - credit-card-data-controller
      summary: Add a new credit card
      operationId: addCardUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: request
          description: request
          required: true
          schema:
            $ref: '#/definitions/CCTransactionRequestModel'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CCGenericResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /charge-amount:
    put:
      tags:
        - credit-card-data-controller
      summary: Charge a credit card with £ currency amount
      operationId: chargeUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CCGenericResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /credit-amount:
    put:
      tags:
        - credit-card-data-controller
      summary: Debit a card with £ currency amount
      operationId: creditUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CCGenericResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /list-cards:
    get:
      tags:
        - credit-card-data-controller
      summary: List all cards from the system
      operationId: getCardsUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CreditCardListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  CCError:
    type: object
    properties:
      message:
        type: string
  CCGenericResponse:
    type: object
    properties:
      balanceRemaining:
        type: string
      cardNumber:
        type: string
      errors:
        $ref: '#/definitions/CCError'
      totalCharge:
        type: string
      totalCredit:
        type: string
  CCTransactionRequestModel:
    type: object
    properties:
      cardNumber:
        type: string
      chargeAmount:
        type: string
      creditAmount:
        type: string
      creditLimit:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      transactionAmount:
        type: string
  CreditCardData:
    type: object
    properties:
      cardNumber:
        type: string
      creditLimit:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      totalCharge:
        type: string
      totalCredit:
        type: string
  CreditCardListResponse:
    type: object
    properties:
      creditCards:
        type: array
        items:
          $ref: '#/definitions/CreditCardData'
      errors:
        $ref: '#/definitions/CCError'